+++ Job-Portal-WebApp/src/main/java/com/hctt/is208/repository/JobPostingDetailRepository.java
@@ -4,7 +4,7 @@ import com.hctt.is208.model.JobPosting;
 import com.hctt.is208.model.JobPostingDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobPostingDetailRepository extends JpaRepository<JobPostingDetail, Long> {
+public interface JobPostingDetailRepository extends JpaRepository<JobPostingDetail, Integer> {
 
 
 }
